requirements.txt
Flask>=2.2
requests>=2.28
pandas>=1.5
plotly>=5.10
python-dotenv>=1.0
